+++ binding.gyp
@@ -4,8 +4,10 @@
      "target_name": "addon",
      "sources": [
        "src/addon.cc",
-       "src/object.cc",
-       "src/async.cc",
+       "src/gray_object.cc",
+       "src/rgb_object.cc",
+       "src/gray_async.cc",
+       "src/rgb_async.cc",
        "src/engine.cc",
        "src/results.cc",
        "src/ccl.cc"
